+++ libsolidity/codegen/CompilerUtils.h
@@ -107,10 +107,11 @@ public:
 	/// and also updates that. For reference types, only copies the data pointer. Fails for
 	/// non-memory-references.
 	/// @param _padToWords if true, adds zeros to pad to multiple of 32 bytes. Array elements
+	/// @param _cleanup if true, adds code to cleanup the value before storing it.
 	/// are always padded (except for byte arrays), regardless of this parameter.
 	/// Stack pre: memory_offset value...
 	/// Stack post: (memory_offset+length)
-	void storeInMemoryDynamic(Type const& _type, bool _padToWords = true);
+	void storeInMemoryDynamic(Type const& _type, bool _padToWords = true, bool _cleanup = true);
 
 	/// Creates code that unpacks the arguments according to their types specified by a vector of TypePointers.
 	/// From memory if @a _fromMemory is true, otherwise from call data.
@@ -309,7 +310,8 @@ private:
 	void cleanHigherOrderBits(IntegerType const& _typeOnStack);
 
 	/// Prepares the given type for storing in memory by shifting it if necessary.
-	unsigned prepareMemoryStore(Type const& _type, bool _padToWords);
+	/// @param _cleanup if true, also cleanup the value when preparing to store it in memory
+	unsigned prepareMemoryStore(Type const& _type, bool _padToWords, bool _cleanup = true);
 	/// Loads type from memory assuming memory offset is on stack top.
 	unsigned loadFromMemoryHelper(Type const& _type, bool _fromCalldata, bool _padToWords);